api: move markdown from lib to client; avoids circular imports
