+++ packages/api/package.json
@@ -24,8 +24,8 @@
     "./client": "./src/client/index.ts",
     "./client/*": "./src/client/*.ts",
     "./dev/*": "./src/dev/*.ts",
-    "./lib/markdown": "./src/lib/markdown/index.ts",
-    "./lib/markdown/*": "./src/lib/markdown/*.ts",
+    "./client/markdown": "./src/client/markdown/index.ts",
+    "./client/markdown/*": "./src/client/markdown/*.ts",
     "./lib/*": "./src/lib/*.ts",
     "./urbit": "./src/urbit/index.ts",
     "./urbit/*": "./src/urbit/*.ts",
