+++ packages/api/src/client/postContent.ts
@@ -6,10 +6,11 @@ import {
   parsePostBlob,
 } from '../client/content-helpers';
 import { assertNever } from '../lib/assertNever';
-import { extractTablesFromContent } from '../lib/markdown/extractTables';
 import { VIDEO_REGEX, containsOnlyEmoji } from '../lib/utils';
 import type { ContentReference } from '../types/references';
 import * as ub from '../urbit';
+import { extractTablesFromContent } from './markdown/extractTables';
+import { convertInlineContent } from './postContentInlines';
 import type { PostContent as ApiPostContent } from './postsApi';
 
 // Inline types
@@ -625,77 +626,10 @@ function convertListing(listing: ub.Listing): ListData {
   }
 }
 
-export function convertInlineContent(inlines: ub.Inline[]): InlineData[] {
-  const nodes: InlineData[] = [];
-  inlines.forEach((inline, i) => {
-    if (typeof inline === 'string') {
-      nodes.push({
-        type: 'text',
-        text: inline,
-      });
-    } else if (ub.isBold(inline)) {
-      nodes.push({
-        type: 'style',
-        style: 'bold',
-        children: convertInlineContent(inline.bold),
-      });
-    } else if (ub.isItalics(inline)) {
-      nodes.push({
-        type: 'style',
-        style: 'italic',
-        children: convertInlineContent(inline.italics),
-      });
-    } else if (ub.isStrikethrough(inline)) {
-      nodes.push({
-        type: 'style',
-        style: 'strikethrough',
-        children: convertInlineContent(inline.strike),
-      });
-    } else if (ub.isInlineCode(inline)) {
-      nodes.push({
-        type: 'style',
-        style: 'code',
-        children: [{ type: 'text', text: inline['inline-code'] }],
-      });
-    } else if (ub.isLink(inline)) {
-      nodes.push({
-        type: 'link',
-        href: inline.link.href,
-        text: inline.link.content ?? inline.link.href,
-      });
-    } else if (ub.isBreak(inline)) {
-      // Most content has a final line break after it -- we don't want to render it.
-      if (i !== inlines.length - 1) {
-        nodes.push({
-          type: 'lineBreak',
-        });
-      }
-    } else if (ub.isShip(inline)) {
-      nodes.push({
-        type: 'mention',
-        contactId: inline.ship,
-      });
-    } else if (ub.isSect(inline)) {
-      nodes.push({
-        type: 'groupMention',
-        group: !inline.sect ? 'all' : inline.sect,
-      });
-    } else if (ub.isTask(inline)) {
-      nodes.push({
-        type: 'task',
-        checked: inline.task.checked,
-        children: convertInlineContent(inline.task.content),
-      });
-    } else {
-      console.warn('Unhandled inline type:', { inline });
-      nodes.push({
-        type: 'text',
-        text: 'Unknown content type',
-      });
-    }
-  });
-  return nodes;
-}
+// Re-exported (and used internally) — definition lives in postContentInlines
+// so that markdown/extractTables can import it without going through
+// postContent (which would create a cycle).
+export { convertInlineContent };
 
 export function prependInline(
   content: BlockData[],
