+++ packages/api/vitest.config.ts
@@ -6,7 +6,7 @@ export default defineConfig({
     include: [
       'src/test/**/*.test.ts',
       'src/__tests__/**/*.test.ts',
-      'src/lib/**/*.test.ts',
+      'src/client/**/*.test.ts',
     ],
     setupFiles: ['./src/test/setup.ts'],
     passWithNoTests: true,
